infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.3.0

